requirements.txt
flask==2.3.3
flask-socketio==5.3.4
python-socketio==5.9.0
python-engineio==4.7.1
websocket-client==1.6.2
